migrate model server install smoke test
