+++ pytest.ini
@@ -3,6 +3,9 @@ testpaths = tests
 
 
 markers =
+    # General
+    polarion: Store polarion test ID
+
     # CI
     smoke: Mark tests as smoke tests
     sanity: Mark tests as sanity tests
